requirements.txt
python-dateutil>=2.5.2
numpy>=1.11
matplotlib>=1.5.2
scipy>=0.18.0
inflect>=0.2.5
libmozdata>=0.1.31
jinja2>=2.8
